+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Changes can also be flagged with a GitHub label for tracking purposes. The URL o
 - Added GPC conditional button option for privacy experiences [#6945](https://github.com/ethyca/fides/pull/6945)
 - Added default identity definitions [#6952](https://github.com/ethyca/fides/pull/6952)
 - Added `display_order` to `PrivacyNotice` and `ExperienceNotices` models [#6939](https://github.com/ethyca/fides/pull/6939)
+- Added new empty message to the action center schema explorer screen [#6972](https://github.com/ethyca/fides/pull/6972)
 
 ### Changed
 - Updated logging configuration to intercept all standard library logs and route them through Loguru[#6891](https://github.com/ethyca/fides/pull/6891)